+++ Generator.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Google.Api.Generator", "Goo
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Google.Api.Generator.Tests", "Google.Api.Generator.Tests\Google.Api.Generator.Tests.csproj", "{7F2B20B6-4F23-49D7-9D12-EC767631E5D8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Google.Api.Generator.Utils", "Google.Api.Generator.Utils\Google.Api.Generator.Utils.csproj", "{5C1198FB-4BF1-4A8A-9CCC-94FF22D57BDC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{7F2B20B6-4F23-49D7-9D12-EC767631E5D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7F2B20B6-4F23-49D7-9D12-EC767631E5D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7F2B20B6-4F23-49D7-9D12-EC767631E5D8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5C1198FB-4BF1-4A8A-9CCC-94FF22D57BDC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5C1198FB-4BF1-4A8A-9CCC-94FF22D57BDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5C1198FB-4BF1-4A8A-9CCC-94FF22D57BDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5C1198FB-4BF1-4A8A-9CCC-94FF22D57BDC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
