+++ Google.Api.Generator.Utils/Formatting/PragmaWarningFormatter.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Google.Api.Generator.RoslynUtils;
+using Google.Api.Generator.Utils.Roslyn;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -21,9 +21,9 @@ using System.Collections.Generic;
 using System.Linq;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
-namespace Google.Api.Generator.Formatting
+namespace Google.Api.Generator.Utils.Formatting
 {
-    internal static class PragmaWarningFormatter
+    public static class PragmaWarningFormatter
     {
         private class PragmaVisitor : CSharpSyntaxRewriter
         {
