+++ Google.Api.Generator.Utils/Roslyn/Keywords.cs
@@ -1,11 +1,22 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Text;
+﻿// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
-namespace Google.Api.Generator.RoslynUtils
+using System.Collections.Immutable;
+
+namespace Google.Api.Generator.Utils.Roslyn
 {
-    internal static class Keywords
+    public static class Keywords
     {
         // Taken from: https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/
         private static IImmutableSet<string> s_keywords = ImmutableHashSet.Create(
