+++ Google.Api.Generator.Utils/Roslyn/ObjectInitExpr.cs
@@ -12,19 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Google.Api.Generator.RoslynUtils
+namespace Google.Api.Generator.Utils.Roslyn
 {
-    class ArgModifier
+    public class ObjectInitExpr
     {
-        public enum Type
-        {
-            Ref,
-            Out,
-        }
-
-        public ArgModifier(Type type, object arg) => (ModType, Arg) = (type, arg);
-
-        public Type ModType { get; }
-        public object Arg { get; }
+        public ObjectInitExpr(string propertyName, object code, bool isDeprecated = false) => (PropertyName, Code, IsDeprecated) = (propertyName, code, isDeprecated);
+        public string PropertyName { get; }
+        public object Code { get; }
+        public bool IsDeprecated { get; }
     }
 }
