+++ Google.Api.Generator.Utils/Typ.cs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Google.Api.Generator.ProtoUtils;
-using Google.Protobuf;
-using Google.Protobuf.Reflection;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System;
 using System.Collections.Generic;
@@ -27,7 +24,7 @@ namespace Google.Api.Generator.Utils
     /// The name "Typ" is delibrately concise as this type is used extensively
     /// throughout the rest of the code.
     /// </summary>
-    internal abstract class Typ : IEquatable<Typ>
+    public abstract class Typ : IEquatable<Typ>
     {
         public sealed class VoidTyp : Typ
         {
@@ -109,30 +106,6 @@ namespace Google.Api.Generator.Utils
             public override string Name => null;
         }
 
-        private static readonly IReadOnlyDictionary<string, Typ> s_wrapperTypes = new Dictionary<string, Typ>
-        {
-            { "google.protobuf.BoolValue", Of<bool?>() },
-            { "google.protobuf.Int32Value", Of<int?>() },
-            { "google.protobuf.UInt32Value", Of<uint?>() },
-            { "google.protobuf.Int64Value", Of<long?>() },
-            { "google.protobuf.UInt64Value", Of<ulong?>() },
-            { "google.protobuf.FloatValue", Of<float?>() },
-            { "google.protobuf.DoubleValue", Of<double?>() },
-            { "google.protobuf.StringValue", Of<string>() },
-            { "google.protobuf.BytesValue", Of<ByteString>() },
-        };
-
-        public static bool IsWrapperType(FieldDescriptor field)
-        {
-            switch (field.FieldType)
-            {
-                case FieldType.Message:
-                    return s_wrapperTypes.ContainsKey(field.MessageType.FullName);
-                default:
-                    return false;
-            }
-        }
-
         public static Typ Void { get; } = new VoidTyp();
 
         public static Typ Of<T>() => Of(typeof(T));
@@ -144,94 +117,6 @@ namespace Google.Api.Generator.Utils
         public static Typ Generic(System.Type genericDef, params Typ[] typeArgs) => new FromGeneric(Of(genericDef), typeArgs);
         public static Typ.GenericParameter GenericParam(string name) => new Typ.GenericParameter(name);
         public static Typ.Special ClassConstraint { get; } = new Special(Special.Type.ClassConstraint);
-
-        public static Typ Of(MessageDescriptor desc)
-        {
-            if (s_wrapperTypes.TryGetValue(desc.FullName, out var wkt))
-            {
-                return wkt;
-            }
-            var ns = desc.File.CSharpNamespace();
-            var isDeprecated = desc.IsDeprecated();
-            var decls = new List<MessageDescriptor>();
-            do
-            {
-                decls.Add(desc);
-                desc = desc.ContainingType;
-            } while (desc != null);
-            decls.Reverse();
-            var typ = Manual(ns, decls[0].Name, isDeprecated: isDeprecated);
-            foreach (var decl in decls.Skip(1))
-            {
-                typ = Nested(Nested(typ, "Types"), decl.Name);
-            }
-            return typ;
-        }
-
-        public static Typ Of(EnumDescriptor desc) => desc.ContainingType == null ?
-            Manual(desc.File.CSharpNamespace(), desc.Name, isEnum: true) :
-            Nested(Nested(Of(desc.ContainingType), "Types"), desc.Name, isEnum: true);
-
-        public static Typ Of(FieldDescriptor desc, bool? forceRepeated = null)
-        {
-            if (desc.IsMap)
-            {
-                if (forceRepeated != null)
-                {
-                    throw new InvalidOperationException("Cannot force repeated on a map field.");
-                }
-                // A map is a repeated message with key and value fields.
-                var kv = desc.MessageType.Fields.InFieldNumberOrder();
-                return Generic(typeof(IDictionary<,>), Of(kv[0]), Of(kv[1]));
-            }
-            // See https://developers.google.com/protocol-buffers/docs/proto3#scalar
-            // Switch cases are ordered as in this doc. Please do not re-order.
-            if (forceRepeated ?? desc.IsRepeated)
-            {
-                switch (desc.FieldType)
-                {
-                    case FieldType.Double: return Of<IEnumerable<double>>();
-                    case FieldType.Float: return Of<IEnumerable<float>>();
-                    case FieldType.Int32: return Of<IEnumerable<int>>();
-                    case FieldType.Int64: return Of<IEnumerable<long>>();
-                    case FieldType.UInt32: return Of<IEnumerable<uint>>();
-                    case FieldType.UInt64: return Of<IEnumerable<ulong>>();
-                    case FieldType.SInt32: return Of<IEnumerable<int>>();
-                    case FieldType.SInt64: return Of<IEnumerable<long>>();
-                    case FieldType.Fixed32: return Of<IEnumerable<uint>>();
-                    case FieldType.Fixed64: return Of<IEnumerable<ulong>>();
-                    case FieldType.SFixed32: return Of<IEnumerable<int>>();
-                    case FieldType.SFixed64: return Of<IEnumerable<long>>();
-                    case FieldType.Bool: return Of<IEnumerable<bool>>();
-                    case FieldType.String: return Of<IEnumerable<string>>();
-                    case FieldType.Bytes: return Of<IEnumerable<ByteString>>();
-                    case FieldType.Message: return Generic(typeof(IEnumerable<>), Of(desc.MessageType));
-                    case FieldType.Enum: return Generic(typeof(IEnumerable<>), Of(desc.EnumType));
-                    default: throw new NotSupportedException($"Cannot get repeated Typ of: {desc.FieldType}");
-                }
-            }
-            switch (desc.FieldType)
-            {
-                case FieldType.Double: return Of<double>();
-                case FieldType.Float: return Of<float>();
-                case FieldType.Int32: return Of<int>();
-                case FieldType.Int64: return Of<long>();
-                case FieldType.UInt32: return Of<uint>();
-                case FieldType.UInt64: return Of<ulong>();
-                case FieldType.SInt32: return Of<int>();
-                case FieldType.SInt64: return Of<long>();
-                case FieldType.Fixed32: return Of<uint>();
-                case FieldType.Fixed64: return Of<ulong>();
-                case FieldType.SFixed32: return Of<int>();
-                case FieldType.SFixed64: return Of<long>();
-                case FieldType.Bool: return Of<bool>();
-                case FieldType.String: return Of<string>();
-                case FieldType.Bytes: return Of<ByteString>();
-                case FieldType.Message: return Of(desc.MessageType);
-                case FieldType.Enum: return Of(desc.EnumType);
-                default: throw new NotSupportedException($"Cannot get Typ of: {desc.FieldType}");
-            }
-        }
 
         /// <summary>The namespace of this typ.</summary>
         public abstract string Namespace { get; }
