+++ Google.Api.Generator/Generation/ServiceBuilderCodeGenerator.cs
@@ -14,7 +14,7 @@
 
 using Google.Api.Gax.Grpc;
 using Google.Api.Gax.Grpc.GrpcCore;
-using Google.Api.Generator.RoslynUtils;
+using Google.Api.Generator.Utils.Roslyn;
 using Google.Api.Generator.Utils;
 using Grpc.Core;
 using Microsoft.CodeAnalysis.CSharp;
@@ -22,8 +22,8 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using static Google.Api.Generator.RoslynUtils.Modifier;
-using static Google.Api.Generator.RoslynUtils.RoslynBuilder;
+using static Google.Api.Generator.Utils.Roslyn.Modifier;
+using static Google.Api.Generator.Utils.Roslyn.RoslynBuilder;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
 namespace Google.Api.Generator.Generation
