+++ Google.Api.Generator/Generation/ServiceMethodGenerator.Signatures.cs
@@ -15,7 +15,7 @@
 using Google.Api.Gax;
 using Google.Api.Gax.Grpc;
 using Google.Api.Generator.ProtoUtils;
-using Google.Api.Generator.RoslynUtils;
+using Google.Api.Generator.Utils.Roslyn;
 using Google.Api.Generator.Utils;
 using Google.Protobuf;
 using Google.Protobuf.Reflection;
@@ -24,8 +24,8 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
-using static Google.Api.Generator.RoslynUtils.Modifier;
-using static Google.Api.Generator.RoslynUtils.RoslynBuilder;
+using static Google.Api.Generator.Utils.Roslyn.Modifier;
+using static Google.Api.Generator.Utils.Roslyn.RoslynBuilder;
 
 namespace Google.Api.Generator.Generation
 {
@@ -189,7 +189,7 @@ namespace Google.Api.Generator.Generation
                                 else
                                 {
                                     // Nested field.
-                                    var code = New(Ctx.Type(Typ.Of(f.Key)))().WithInitializer(NestInit(f, ofs + 1).ToArray());
+                                    var code = New(Ctx.Type(ProtoTyp.Of(f.Key)))().WithInitializer(NestInit(f, ofs + 1).ToArray());
                                     yield return new ObjectInitExpr(f.Key.CSharpPropertyName(), code);
                                 }
                             }
